test code commit

+++ dy-java-web/src/main/java/com/dyj/web/domain/vo/AccessTokenVo.java
@@ -14,7 +14,7 @@ public class AccessTokenVo extends BaseVo{
     /**
      * access_token接口调用凭证超时时间，单位（秒)
      */
-    private Integer expires_in;
+    private Long expires_in;
 
     /**
      * 日志ID
@@ -29,7 +29,7 @@ public class AccessTokenVo extends BaseVo{
     /**
      * refresh_token凭证超时时间，单位（秒)
      */
-    private Integer refresh_expires_in;
+    private Long refresh_expires_in;
     /**
      * 用户刷新access_token
      */
@@ -51,11 +51,11 @@ public class AccessTokenVo extends BaseVo{
         this.access_token = access_token;
     }
 
-    public Integer getExpires_in() {
+    public Long getExpires_in() {
         return expires_in;
     }
 
-    public void setExpires_in(Integer expires_in) {
+    public void setExpires_in(Long expires_in) {
         this.expires_in = expires_in;
     }
 
@@ -75,11 +75,11 @@ public class AccessTokenVo extends BaseVo{
         this.open_id = open_id;
     }
 
-    public Integer getRefresh_expires_in() {
+    public Long getRefresh_expires_in() {
         return refresh_expires_in;
     }
 
-    public void setRefresh_expires_in(Integer refresh_expires_in) {
+    public void setRefresh_expires_in(Long refresh_expires_in) {
         this.refresh_expires_in = refresh_expires_in;
     }
